requirements.txt
onnxruntime==1.19.2
numpy==1.26.4
Pillow==10.4.0
opencv-python==4.10.0.84
gradio==4.44.0
requests==2.32.3
tqdm==4.66.5

